+++ modules/mcp_sentinel_graphql/tests/src/Kernel/GraphqlGovernanceSubscriberTest.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Drupal\Tests\mcp_sentinel_graphql\Kernel;
 
+use Drupal\Tests\mcp_sentinel\Traits\McpAuditSchemaTestTrait;
 use Drupal\Core\Cache\CacheBackendInterface;
 use Drupal\Core\Config\ConfigCrudEvent;
 use Drupal\graphql\Event\OperationEvent;
@@ -41,6 +42,8 @@ use Symfony\Component\HttpFoundation\Request;
 #[RunTestsInSeparateProcesses]
 final class GraphqlGovernanceSubscriberTest extends KernelTestBase {
 
+  use McpAuditSchemaTestTrait;
+
   use UserCreationTrait;
 
   /**
@@ -74,7 +77,7 @@ final class GraphqlGovernanceSubscriberTest extends KernelTestBase {
   protected function setUp(): void {
     parent::setUp();
     $this->installEntitySchema('user');
-    $this->installSchema('audit_chain', ['audit_chain_log']);
+    $this->installAuditChainSchema();
     $this->installSchema('mcp_sentinel', [
       'mcp_sentinel_content_locks',
     ]);