+++ tests/src/Kernel/McpClassificationEmptyMapRegressionTest.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Drupal\Tests\mcp_sentinel\Kernel;
 
+use Drupal\Tests\mcp_sentinel\Traits\McpAuditSchemaTestTrait;
 use Drupal\Core\Access\AccessResultReasonInterface;
 use Drupal\KernelTests\KernelTestBase;
 use Drupal\mcp_sentinel\Controller\McpContextController;
@@ -41,6 +42,8 @@ use Symfony\Component\HttpKernel\HttpKernelInterface;
 #[RunTestsInSeparateProcesses]
 final class McpClassificationEmptyMapRegressionTest extends KernelTestBase {
 
+  use McpAuditSchemaTestTrait;
+
   use UserCreationTrait;
   use McpClassificationTestTrait;
 
@@ -74,7 +77,7 @@ final class McpClassificationEmptyMapRegressionTest extends KernelTestBase {
     $this->installEntitySchema('taxonomy_term');
     $this->installEntitySchema('path_alias');
     $this->installSchema('system', ['sequences']);
-    $this->installSchema('audit_chain', ['audit_chain_log']);
+    $this->installAuditChainSchema();
     $this->installSchema('mcp_sentinel', ['mcp_sentinel_content_locks']);
     $this->installConfig(['filter', 'node', 'mcp_sentinel', 'user']);
     require_once \Drupal::root() . '/' . \Drupal::service('extension.list.module')