+++ tests/src/Kernel/McpDlpPathBehaviourTest.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Drupal\Tests\mcp_sentinel\Kernel;
 
+use Drupal\Tests\mcp_sentinel\Traits\McpAuditSchemaTestTrait;
 use Drupal\Core\Access\AccessResultReasonInterface;
 use Drupal\KernelTests\KernelTestBase;
 use Drupal\mcp_sentinel\Entity\McpPolicyProfile;
@@ -33,6 +34,8 @@ use Symfony\Component\HttpKernel\HttpKernelInterface;
 #[Group('mcp_sentinel')]
 #[RunTestsInSeparateProcesses]
 final class McpDlpPathBehaviourTest extends KernelTestBase {
+
+  use McpAuditSchemaTestTrait;
 
   use UserCreationTrait;
 
@@ -73,7 +76,7 @@ final class McpDlpPathBehaviourTest extends KernelTestBase {
     parent::setUp();
     $this->installEntitySchema('user');
     $this->installSchema('system', ['sequences']);
-    $this->installSchema('audit_chain', ['audit_chain_log']);
+    $this->installAuditChainSchema();
     $this->installSchema('mcp_sentinel', ['mcp_sentinel_content_locks']);
     $this->installConfig(['system', 'user', 'mcp_sentinel']);