+++ tests/src/Kernel/McpRawSqlCommandTest.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Drupal\Tests\mcp_sentinel\Kernel;
 
+use Drupal\Tests\mcp_sentinel\Traits\McpAuditSchemaTestTrait;
 use Drupal\KernelTests\KernelTestBase;
 use Drupal\mcp_sentinel\Drush\Commands\McpSentinelSqlCommands;
 use Drupal\node\Entity\Node;
@@ -34,6 +35,8 @@ use Symfony\Component\Console\Output\BufferedOutput;
 #[Group('mcp_sentinel')]
 #[RunTestsInSeparateProcesses]
 final class McpRawSqlCommandTest extends KernelTestBase {
+
+  use McpAuditSchemaTestTrait;
 
   /**
    * {@inheritdoc}
@@ -76,7 +79,7 @@ final class McpRawSqlCommandTest extends KernelTestBase {
   protected function setUp(): void {
     parent::setUp();
 
-    $this->installSchema('audit_chain', ['audit_chain_log']);
+    $this->installAuditChainSchema();
     $this->installEntitySchema('user');
     $this->installEntitySchema('node');
     $this->installConfig(['audit_chain', 'mcp_sentinel']);